Added casts to make compiler happy
git-svn-id: https://svn.aros.org/svn/aros/trunk@19012 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ workbench/c/Version.c
@@ -385,9 +385,9 @@ int makekickver()
 
     /* Fill in struct parsedver. */
     ArosInquire(
-                AI_ArosVersion, &parsedver.version,
-                AI_ArosReleaseMinor, &parsedver.revision,
-                AI_ArosReleaseDate, &parsedver.days,
+                AI_ArosVersion, (ULONG)&parsedver.version,
+                AI_ArosReleaseMinor, (ULONG)&parsedver.revision,
+                AI_ArosReleaseDate, (ULONG)&parsedver.days,
                 TAG_DONE);
 
 #define KICKSTRLEN 5
